vtorc: move cooldown logic to standalone files for fork maintainability
Move all cooldown types, functions, metric, and tests into dedicated
files (topology_recovery_cooldown.go and topology_recovery_cooldown_test.go)
so that upstream backports only conflict on the two small call sites in
topology_recovery.go (~19 lines) rather than the full implementation.
