+++ go/vt/vtorc/logic/topology_recovery.go
@@ -22,7 +22,6 @@ import (
 	"errors"
 	"fmt"
 	"math/rand/v2"
-	"os"
 	"sync/atomic"
 	"time"
 
@@ -83,9 +82,6 @@ var (
 
 	// recoveriesFailureCounter counts the number of failed recoveries that VTOrc has performed
 	recoveriesFailureCounter = stats.NewCountersWithMultiLabels("FailedRecoveries", "Count of the different failed recoveries performed", recoveriesCounterLabels)
-
-	// recoveriesCooldownSkipped counts the number of cluster-wide recoveries skipped due to shard cooldown
-	recoveriesCooldownSkipped = stats.NewCountersWithMultiLabels("RecoveriesCooldownSkipped", "Count of cluster-wide recoveries skipped due to shard cooldown", recoveriesCounterLabels)
 
 	// shardLockTimings measures the timing of LockShard operations.
 	shardLockTimingsActions = []string{"Lock", "Unlock"}
@@ -228,74 +224,6 @@ func resolveRecovery(topologyRecovery *TopologyRecovery, successorInstance *inst
 		topologyRecovery.IsSuccessful = true
 	}
 	return writeResolveRecovery(topologyRecovery)
-}
-
-const recoveryCooldownKeyPrefix = "recovery_cooldown"
-
-type recoveryCooldownMarker struct {
-	Timestamp    time.Time `json:"timestamp"`
-	AnalysisCode string    `json:"analysis_code"`
-	RecoveryName string    `json:"recovery_name"`
-	InstanceID   string    `json:"instance_id"`
-}
-
-func recoveryCooldownKey(keyspace, shard string) string {
-	return fmt.Sprintf("%s/%s/%s", recoveryCooldownKeyPrefix, keyspace, shard)
-}
-
-// checkRecoveryCooldown returns true if a cooldown is active for the shard.
-func checkRecoveryCooldown(ctx context.Context, keyspace, shard string) (bool, error) {
-	cooldownDuration := config.GetRecoveryCooldownDuration()
-	if cooldownDuration <= 0 {
-		return false, nil
-	}
-
-	key := recoveryCooldownKey(keyspace, shard)
-	val, err := ts.GetSingleMetadata(ctx, key)
-	if err != nil {
-		return false, fmt.Errorf("failed to check recovery cooldown for %s/%s: %w", keyspace, shard, err)
-	}
-	if val == "" {
-		return false, nil
-	}
-
-	var marker recoveryCooldownMarker
-	if err := json.Unmarshal([]byte(val), &marker); err != nil {
-		log.Warningf("Corrupted recovery cooldown marker for %s/%s, ignoring: %v", keyspace, shard, err)
-		return false, nil
-	}
-
-	if time.Since(marker.Timestamp) >= cooldownDuration {
-		return false, nil
-	}
-
-	return true, nil
-}
-
-// writeRecoveryCooldown writes a cooldown marker after a successful cluster-wide recovery.
-func writeRecoveryCooldown(keyspace, shard, analysisCode, recoveryName string) {
-	if config.GetRecoveryCooldownDuration() <= 0 {
-		return
-	}
-
-	hostname, _ := os.Hostname()
-	marker := recoveryCooldownMarker{
-		Timestamp:    time.Now(),
-		AnalysisCode: analysisCode,
-		RecoveryName: recoveryName,
-		InstanceID:   hostname,
-	}
-	data, err := json.Marshal(marker)
-	if err != nil {
-		log.Errorf("Failed to marshal recovery cooldown marker: %v", err)
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	if err := ts.UpsertMetadata(ctx, recoveryCooldownKey(keyspace, shard), string(data)); err != nil {
-		log.Errorf("Failed to write recovery cooldown marker for %s/%s: %v", keyspace, shard, err)
-	}
 }
 
 // recoverPrimaryHasPrimary resets the replication on the primary instance
